+++ report.php
@@ -46,7 +46,7 @@ if ($mode == "list") {
                 <h2><span class="title"><?php echo $report->getBestIdentifier(); ?></span></h2>
                 <div id="report-button">
                     <a class="report-button" id="update-report-button" name="<?php echo $collectionId; ?>" href="#">run update</a>
-                    <a class="report-button" href="./report.php?id=<?php echo $collectionId; ?>&mode=list">download data</a>
+                    <a class="report-button" href="/csv/<?php echo $collectionId; ?>">download data</a>
                 </div>
                 <div id="report-info">
                     <span class="badge artifacts-count"><span class="num"><?php echo $report->getArtifactsCount(); ?></span> artifacts;</span>
@@ -58,11 +58,11 @@ if ($mode == "list") {
     
                 <div id="share">    
                     <!-- based on code here: https://dev.twitter.com/docs/tweet-button -->    
-                    <span id="permalink"><span class="label"> Permalink: </span><a class="copyable", href="<?php echo "http://" . $_SERVER['HTTP_HOST'] . $_SERVER['SCRIPT_NAME'] . "?id=" . $collectionId; ?>"><?php echo "http://" . $_SERVER['HTTP_HOST'] . $_SERVER['SCRIPT_NAME'] . "?id=" . $collectionId; ?></a><a href="#" id="copy-permalink">copy</a></span>
+                    <span id="permalink"><span class="label"> Permalink: </span><a class="copyable", href="<?php echo "http://" . $_SERVER['HTTP_HOST'] . "/collection/" . $collectionId; ?>"><?php echo "http://" . $_SERVER['HTTP_HOST'] . "/collection/" . $collectionId; ?></a><a href="#" id="copy-permalink">copy</a></span>
                     <script src="//platform.twitter.com/widgets.js" type="text/javascript"></script>    
                     <span class="tweet-this">    
                         <a href="https://twitter.com/share" class="twitter-share-button"    
-                           data-url="<?php echo "http://total-Impact.org/report.php?id=" . $collectionId ?>"
+                           data-url="<?php echo "http://total-impact.org/report.php?id=" . $collectionId ?>"
                            data-via="mytotalImpact"    
                            data-text="<?php echo "Check out the total-Impact of " . $report->getBestIdentifier() . ": "; ?>"
                            data-count="horizontal">Tweet</a>